+++ lib/traps/__init__.py
@@ -4,9 +4,10 @@ from .QTrapOverlay import QTrapOverlay
 from .QTrapMenu import QTrapMenu
 from .commands import (
     AddTrapCommand, RemoveTrapCommand,
-    MoveCommand, RotateCommand, WheelCommand)
+    MoveCommand, RotateCommand, WheelCommand, LockCommand)
 
 
 __all__ = ('QTrap QTrapGroup QTrapOverlay QTrapMenu '
            'AddTrapCommand RemoveTrapCommand '
-           'MoveCommand RotateCommand WheelCommand').split()
+           'MoveCommand RotateCommand WheelCommand '
+           'LockCommand').split()